Refined places where askOption need to be used.
- Make sure that symbols are managed only within the Presentation Compiler
  Thread.
- Prevent NullPointerException during
  SemanticHighlightingReconciliation.createSemanticDecorationManagers (as the
  editor may be null).

Also, in SymbolClassification, there is no need to store the source file
content in a string.  This is really needed only if the user enabled
`useSyntacticHints`. Not doing it by default can save a few cycles,
particularly for big files.

WARNING!!

There is a problem with this commit, which is the tests can no longer be run.
We definitely need a solution for this!